# docs/swagger.yaml
definitions:
  handler.getMealsResponse:
    description: 食事を追加する
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: integer
      mealtype:
        type: string
      memo:
        type: string
      protein:
        type: number
    type: object
  handler.postMealRequest:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      mealtype:
        type: string
      memo:
        type: string
      protein:
        type: number
    type: object
  handler.postMealResponse:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: integer
      mealtype:
        type: string
      memo:
        type: string
      protein:
        type: number
    type: object
  handler.putMealsRequest:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      mealtype:
        type: string
      memo:
        type: string
      protein:
        type: number
    type: object
  handler.putMealsResponse:
    properties:
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      id:
        type: integer
      mealtype:
        type: string
      memo:
        type: string
      protein:
        type: number
    type: object
  handler.userLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.userLoginResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
  handler.userSignupRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  handler.userSignupResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: API server for the Nutrition App
  title: Nutrition App API
  version: "1"
paths:
  /meals:
    get:
      consumes:
      - application/json
      description: 食事を追加する
      operationId: getMeals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/handler.getMealsResponse'
              type: array
            type: array
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
    post:
      consumes:
      - application/json
      description: 食事を追加する
      operationId: postMeals
      parameters:
      - description: 食事情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.postMealRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.postMealResponse'
        "400":
          description: ""
        "401":
          description: ""
        "500":
          description: ""
  /meals/{id}:
    delete:
      consumes:
      - application/json
      description: 食事を消去する
      operationId: deleteMealsId
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
    get:
      consumes:
      - application/json
      description: 食事を追加する
      operationId: getMealsId
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getMealsResponse'
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
    put:
      consumes:
      - application/json
      description: 食事を追加する
      operationId: putMealsId
      parameters:
      - description: 食事情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.putMealsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.putMealsResponse'
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
  /users/lgin:
    post:
      consumes:
      - application/json
      description: ログイン
      operationId: postUsersLogin
      parameters:
      - description: ログイン情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.userLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userLoginResponse'
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
  /users/signup:
    post:
      consumes:
      - application/json
      description: サインアップ
      operationId: postUsersSignup
      parameters:
      - description: サインイン情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.userSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userSignupResponse'
        "400":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
swagger: "2.0"